Catch BrokenPipeError, always check all clusters
BrokenPipeError can be caused by a broken UGE installation.
A possible underlying error is eg.
   OSError: /dls_sw/cluster/GE/UGE8.6.14/lib/lx-amd64/libdrmaa.so:
   cannot open shared object file: No such file or directory
but this error is happening in a separate process and shows up as
BrokenPipeError.

If this happens (or any other persistent error) then close the
connection to the failed cluster subprocess and intercept further calls
and show a warning instead. This will reduce the log spam slightly.