Import bitshuffle before omptbx
bitshuffle must be loaded before omptbx gets a chance to introduce its
very helpful and deadly code into the library space. If this is not done
here and omptbx is allowed to load first then on using bitshuffle omptbx
will immediately crash the python interpreter hard, with a useful

  omptbx error: deadlock in using lock variable

message, despite not having been asked for an opinion on the matter.

Fixes #5