Workaround autoPROC/20200918 beam centre units bug
We reported the direct beam position in pixels within that
ISPyB-compatible XML file, when other programs apparently
reported it in mm. So we wanted to change it to mm as well,
but did that fix in two places at the same time ...
resulting in multiplying by (pixelsize)**2 instead of just
pixelsize.