Exponential backoff for multiplex trigger
In cases where a xia2-dials job for the current data collection
completes before the xia2-dials job for a previous data collection
this can result in either multiplex not being triggered at all, or
triggered, but not including as many previous data collections as
the user may expect. Instead check the processing status of the
registered xia2-dials jobs for related data collections, and if any
are "running" or "queued" then checkpoint the current message with
a time delay.

Use an exponential backoff to avoid overloading the service.
The backoff can be controlled using the parameters backoff-delay
and backoff-multiplier, where the calculated delay is
backoff-delay * backoff-multiplier ** ntry
The message will be resent for a maximum of backoff-max-try
attempts, at which point the trigger will continue and trigger a
multiplex job with whichever xia2-dials results are available.