Dummy insert particle picking for ISPyB service (#48)
Add a dummy `do_insert_particle_picker` method. This just logs a message and returns a success if called. This should stop potential future crashes when the Relion wrapper will be sending messages for insertion into the `ParticlePicker` table. Also reformat the dummy methods for 2D and 3D classification stages.