Remove use_sample_id for Therm_6 definition
I don't think the sample_id is relevant since it links to a thermolysin
Protein entry (and this is apparently thaumatin, despite the name) which
means that any downstream dimple/mrbump jobs fail anyway. A side effect
of linking to an existing sample_id is that we end up triggering a
xia2.multiplex job with 951 (and counting) prior datasets associated
with this sample.