Add copy_sample_id option to dc_sim
This will (partially) copy an existing BLSample entry and use that
for the subsequent dc_sim data collection, rather than adding
infinitely more data collections to an existing sample.
The crystalId is copied across, but not the containerId (as I don't
know how nicely this will play with e.g. SynchWeb).

See also #49