Use selectinload instead of joinedload
joinedload can be inefficient for large collections as every column
of the parent object is repeated for every row of the collection.
Using selectin load appears to be much more efficient in this case,
as the number of data collections associated with a single
ProcessingJob can potentially be large (many hundreds) for a
xia2.multiplex job.

See also:
  https://docs.sqlalchemy.org/en/14/orm/loading_relationships.html#what-kind-of-loading-to-use