Services to bridge between ActiveMQ and RabbitMQ (#82)
Service DLS Bridge takes messages from ActiveMQ and moves them to RabbitMQ. DLS Reverse Bridge does the opposite. Which queues to take from and move to are specified by a member dictionary (`{source: destination}`) called `queues`. At the moment this is configured to take from `transient.transaction` on ActiveMQ to `transient.transaction` on RabbitMQ and `transient.destination` in the other direction.

The credentials file the services use to connect to the message broker they are not running on is hardcoded at the moment so will always be the "live" version.