Use Mesh3D dcg.experimentType
The first dc of a 3d grid scan may not have any related dcids
as the second dc may not have been registered by gda at the time
the we are processing the first grid scan. Therefore use the
presence/absence of the experimentType=Mesh3D in addition to
presence of dgc_dcids to decide on whether to treat as 2D or 3D
grid scan.