Remove potentially expensive leftover flag
Back in the day the controller service used JMS Advisory messages to
check queue sizes. This was weirdly slow and resource-intensive on the
ActiveMQ side, and in SCI-7581 we moved to a better mechanism to achieve
the same thing.

However, we left the 'transformation' flag in place, which attaches some
ActiveMQ plugin to the queue. Loading this plugin likely is also
expensive on the server side, and has been implicated in today's
ActiveHiccup.

Since we don't use advisory messages any more, just remove the
translate-advisory-to-json flag, this should stop ActiveMQ from loading
the plugin, which may improve things, but will certainly not make them
worse.