Remove ActiveMQ logging (#108)
This was introduced back in the day before Graylog to provide central
logging, and was retained after Graylog was introduced as it was more
reliable. Since then Graylog has become a lot more stable and we now no
longer need it.

There are no plans to support this on RabbitMQ, so it's now time to
retire our home-grown log distribution system.