Add checks on RelativeIceThickness results to EM DC sim (#118)
The Relion pipeline now includes the calculation of a five figure summary of "relative ice thickness" for each micrograph with these results ending up in the `RelativeIceThickness` ISPyB table. Check that that is all working ok within the daily simulated data collection. Check that the number of records is one per micrograph and that the values are reasonable.