Improve metric integrations in xrc and trigger
This provides an improvement on the way prometheus metrics work within the xrc and trigger services. The prometheus initialistions are done within the utility prometheus_metrics "dlstbx.util.prometheus_metrics". Only the metric created needs to be done within the required service. With in the service initialization set the service level metrics if they are desired, indicated by a '-m' flag at launch. If not set service metrics to a blank metrics class.

mvp, util needs improvement.