Update dimple wrapper for procrunner timeout handling
procrunner.run, when used with timeout= should also specify
raise_timeout_exception=True, which will cause it to raise
subprocess.TimeoutExpired() if the timeout is exceeded.

(Previously the return object would have a 'timeout' field, but that is
going away)

Also: no longer use array access on the return object as that is also
going away.