Trigger i19-2 EIGER PIA at END of data collection
The meta.h5 file doesn't appear to be written until the end of data
collection, and FormatNXmx doesn't like this, as it encounters a
KeyError on attempting to access nx_detector.saturation_value
which contains an external link to the meta.h5 file.

As a temporary workaround, trigger PIA at the END of data collection
in the hope that the meta.h5 file has been written by this time.

Also satisfy some mypy typing complaints.