Use loaded service zocalo config instance
Reading the zocalo configuration on every request, as was currently
implemented, results in the graylog plugin being re-activated with
every request, with the end result that we're adding a new graylog
handler to the logger with every request, thus duplicating messages
many times over in graylog.

Instead, pass down the loaded zocalo configuration instance from
the service.