Downgrade some pika error messages to warn level
so they will still show up in logging, but are closer to our expected
semantics, ie. warnings demand attention but are recoverable.

A note on logging filters: If attached to a logger they will only fire
on that specific logger, not any loggers underneath. Hence the need to
attach this one to two specific loggers rather than to 'pika' directly.