New SSXPlotter service
Watches a given pia.json (as written by the jsonlines service)
and when all expected results have appeared it will generate
a png (and *t.png). Intended to be saved in ISPyB as attachments
for display in SynchWeb.

It will watch for results for {timeout} seconds and if not all
results have appeared by that time it will generate a plot with
results have appeared.