Only load the models for images being analysed
For nexus files setting load_models=True (i.e. the default) when calling
ExperimentListFactory.from_filenames() constructs an experiment object for
every experiment in the data set. For an ssx dose series this could mean
constructing over 250k experiments in order to do spotfinding on a single
image, resulting in an approximately 5-fold increase in the time taken to
do spotfinding. This change sets  for nexus files and
then explicitly calls expt.load_models() for only the images being
analysed.