Add dcg functionality to custom data dc_sim
The dc_sim method now accepts a supplied dcg ID as input and has an
option to make the simulated data collection as a data collection group.
If a dcg is supplied, the default behaviour is to output as a dcg. If
one is not supplied, the default is to output as a single data
collection.

The "dcg" parameter in the defined scenarios has been renamed to match
the "is_dcg" parameter that can be supplied as a command line input,
which was done to make the bool nature of the input parameter more
obvious.