Filewatcher: Explicitly log filenames of long stats
We frequently face a task of tracking down backwards from a long-stat-time
alert to what the actual files being read were. This updates the Filewatcher
to explicitly log the stat if it takes "a long time", defined as the
threshold that we currently alert on (5s).

Also, fix a bug where HDF5 read time was included in the stat-time.

The filewatcher service instance of _Profiler has been reworked to be
slightly friendlier and more limited in it's functionality.