Consolidate looped DC lookups
For ispyb_related_sweeps and ispyb_images, there were looped lookups over every
DataCollection in the group in order to extract a couple of parameters. This
caused issues for groups with a large number of collections because it was
pulling down very large amounts of data over lots of queries (2300 queries for
12MB of SQL observed for 400 collections!).

This replaces these with a single query getting only the required information.