Add Metal ID postprocessing pipeline (#240)
Add metal id postprocessing pipeline

Add the functionality to process metal identification experiments whereby two data collections are taken; one above and one below the absorption edge of an element. The processing takes anomalous fourier maps (obtained from the SHELX AnoDe package) of both data collections and subtracts one from the other to produce a "double difference map" that should reveal the location of atoms/ions of the given element within the protein structure.

The workflow for this requires two mtz files from datasets recorded from the same crystal above and below the absorption edge of an anomalous scatterer and a pdb file for native phasing of the molecule. The overall scale of the datasets are matched using scaleit, before dimple is run on both of them. The anode.pha output from both runs of dimple are then used to create a double difference map by subtracting the below data from the above data, which should give a map corresponding to the locations of the anomalous scatterers. The following additions/changes have been made to achieve this:

- Add scaleit wrapper for adjusting the overall scale of two single dataset mtz files relative to each other. 

- Allow suppression of sending of dimple results to ispyb dimple wrapper. This lets the metal_id workflow use the dimple wrapper without having to upload the dimple results to ISPyB.

- Add metal_id wrapper. This takes two anode.pha files and two corresponding pdb files. It checks that the two pdb are suitably similar before creating a double difference map from the two anomalous Fourier maps. It then finds and renders images of peaks in the map.

- Add trigger_metal_id function to trigger service.