Restart pod if slurm api call fails
If Slurm token expires, the pod slurm api calls will fail but the
service will handle it and keep running. The k8s pod running the service
requires a restart to pick up the new token. This Re-raises the error
after logging it to cause the pod to crash and restart it and pick up
the new Slurm token. This also puts the service in a crashloop in the
case of persistent problems, which would be caught by alerts.