Fix identical dataset name bug in scaleit wrapper
If datasets have identical crystal name, dataset name and unit cell
parameters, CAD will keep them in the same dataset in the MTZ file. This
causes Scaleit to get confused with its auto-column assignment. This fix
forces the two reference columns added to the data to be scaled to be
named differently from the main dataset.