Add ligand fitting quality threshold to determine pipeline output behaviour
`send_attachments_to_ispyb` function updated such that it will only send the final output files if the CC quality threshold is met. If the CC threshold is not met then only pipeline.log is sent to ISPyB. Synchweb will now display a tick only if the pipeline is successful in fitting a ligand of sufficient quality.