Add default value to pydantic fields (#266)
PR #247 updated dlstbx to pydantic 2. Some fixes were applied in #265. Despite this, it was found that the multiplex trigger was failing validation in production. The cause was found to be due to some optional fields missing a default value, causing pydantic to fail validation as without a default value, pydantic still expects the field to exist in the input message but will allow its value to be None. In this fix, the remaining pydantic fields that do not have specified default values have been given default values of None .