Fix pydantic 2 validator
A change in pydantic 2 means that in some circumstances, the model
validator is passed the model itself, instead of a dictionary of values
via the values argument. This is now the case in the ssx-plotter.
An assert statement has been added to catch if something other than the
model is passed as the values argument.