Move logic from trigger to wrapper
Instead of having logic for finding mtz files from clusters in the
trigger service, this is done in the multiplex wrapper itself. The
mtz files are then sent to the trigger function via the zocalo message,
which is already incorporated into the parameter map if the message is
of type dict.