Fix file copying and synchweb display
Blank integration needs to be uploaded at the start of the recipe,
otherwise processing appears in the downstream processing on snychweb.
Changes made allow a pre-registered integration id to be used when
inserting integration data for the 'All data' dataset and to create a
new integration id for the cluster data.

Applied fixes to file copying/uploading. Cluster data was being ignored
because code required a complete match with file patterns. Now it just
requires the file to end with the file pattern.