Trigger mrbump and big ep for each trigger
Adds scaled_unmerged data to the environment and generalises the names
of the file paths stored in the environment as they are not used by
just one downstream pipeline. This changes requires changes to the
recipe to pass placholders matching the name of the environment key,
preceded by a $ symbol.