Change file keeping logic
Fix a bug in the logic to stop the wrong file being uploaded when
destination was not set appropriately. Upload files that are not on the
final result list. These are files that are not as important long term
and will be removed from disk after 1 year but are still useful to have
available on disk in the short term.

Fix a bug where the primary log files that are to be listed first were
being uploaded twice for the main data.

This logic is not complete as it doesn't handle files of the same name
that exist in the main directory as well as the cluster directory. Also
still need to decide which, if any, files from the main directory that
should be uploaded to ispyb for each clusters results.