Track dlstbx_align_crystal processing status
Removes the initial creation of the screening record from the wrapper.
This will be performed in a separate recipe step along with the creation
of an AutoProcProgram record, which will be linked to the screening
record.

The AutoProcProgram record will be updated to reflect the processing
status of the pipeline. Two conditional ispyb commands have been added
to the wrapper to update the processing status accordingly when the
pipeline completes showing success if a solution is found and failure
if no valid solution exists.

If any of the steps in the pipeline fail with a non-zero return code,
the processing status will be set to 'Failed' via a separate recipe
step.