Set xoalign wrapper up for process status tracking
Also removes beamline check from within the wrapper as this will cause
the wrapper to succeed but when it hasn't really run. In the workflow
a screening record will have already been created. If beamline
restriction is still desired, this should be done in the trigger step
and not in the wrapper.