Search for matching processing with appid
Previous logic looked for same proc_prog run on the matching dcids.
However, if multiple runs of the same proc_prog ran on the same dcids,
the mtz files may both get picked up from the same dcid. Instead, use
autoProcProgramId of the job that sent the trigger message and find
processing on the matching dcid that ran with the same processing
environment as well as proc_prog.