Use pipeline_final and clean up file uploading
Use of pipeline final set by the recipe "pipeline_final" job parameter.

Primary results files to copy to pipeline_final and to upload to ISPyB
set in the recipe too, via the "primary_result_files" job parameter.

Logic set up to optionally use the pipeline_final if it is defined.
Rendered images will be copied to piepline final but will not be
uploaded as an attachment. They get uploaded to the MXMRRunBlob table
instead.