Remove lb from list of industrial codes
This is not an industrial code, it's an academic XChem proposal. This
was marked as industrial in the past as a quick way of preventing
phasing pipelines, since XChem run their own stack. For OpenBIND there
are lb visits that run on other beamlines, and they would like to run
the phasing pipelines.

This enables everywhere for now, and we can scope it to non-I04-1 later.